requirements.txt
aiogram @ git+https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
openai==0.27.8
python-dotenv==1.0.0
asyncio==3.4.3